requirements.txt
# Core
fastapi==0.109.2
uvicorn==0.27.1
flask==2.3.3
python-multipart==0.0.9

# Database
asyncpg==0.29.0        # Async PostgreSQL driver
psycopg2==2.9.9        # Sync PostgreSQL adapter
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment
python-dotenv==1.0.1

# Validation
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

# Optional
pytest==8.1.1
httpx==0.27.0